Locker contextvars (#48)
* Delete cache entries when functions fail (#42)

* test with postgresql

* fix locker bug

* cleanup

* simplify

* comment

* fix env

* 2-steps return

* cleanup

* safer unlocking

* fix flaky tests

* change wait

* tune

* make SETTINGS and LAST_PRIMARY_KEYS contextvars (#44)

* make LAST_PRIMARY_KEYS a contextvar

* cleanup

* make settings contextvar

* better name for contextvars

* restore concurrency test

* wip: fix threading for settings

* implement context

* better locker

* introduce __context__

* add pydantic model

* handle settings through pydantic

* tune

* add dotenv to setup

* cleanup

* better docstring

* typo

* specify case sensitive

* cleanup

* cleanup

* better settings

* template update

* fix typing

* init settings

* contextvars on top

* fix reset

* do not expand args and kwargs

* rename func

+++ cacholote/cache.py
@@ -15,12 +15,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import contextvars
 import datetime
 import functools
 import json
 import time
 import warnings
-from typing import Any, Callable, Dict, TypeVar, Union, cast
+from typing import Any, Callable, Dict, Optional, TypeVar, Union, cast
 
 import sqlalchemy
 import sqlalchemy.orm
@@ -29,37 +30,38 @@ from . import clean, config, decode, encode, utils
 
 F = TypeVar("F", bound=Callable[..., Any])
 
-LAST_PRIMARY_KEYS: Dict[str, Any] = {}
+LAST_PRIMARY_KEYS: contextvars.ContextVar[Dict[str, Any]] = contextvars.ContextVar(
+    "cacholote_last_primary_keys"
+)
 
-_LOCKER = None
+_LOCKER = "__locked__"
 
 
 def _update_last_primary_keys_and_return(
-    session: sqlalchemy.orm.Session, cache_entry: Any
+    session: sqlalchemy.orm.Session, cache_entry: Any, tag: Optional[str]
 ) -> Any:
-    if _LOCKER:
-        # Wait until unlocked
-        warned = False
-        while cache_entry.result == _LOCKER:
-            session.refresh(cache_entry)
-            if not warned:
-                warnings.warn(
-                    f"can NOT proceed until the cache entry is unlocked: {cache_entry!r}."
-                )
-                warned = True
-            time.sleep(1)
+    # Wait until unlocked
+    warned = False
+    while cache_entry.result == _LOCKER:
+        session.refresh(cache_entry)
+        if not warned:
+            warnings.warn(
+                f"can NOT proceed until the cache entry is unlocked: {cache_entry!r}."
+            )
+            warned = True
+        time.sleep(1)
     # Get result
     result = decode.loads(cache_entry._result_as_string)
     cache_entry.counter += 1
-    if config.SETTINGS["tag"] is not None:
-        cache_entry.tag = config.SETTINGS["tag"]
+    if tag is not None:
+        cache_entry.tag = tag
     session.commit()
-    LAST_PRIMARY_KEYS.update(cache_entry._primary_keys)
+    LAST_PRIMARY_KEYS.set(cache_entry._primary_keys)
     return result
 
 
 def _clear_last_primary_keys_and_return(result: Any) -> Any:
-    LAST_PRIMARY_KEYS.clear()
+    LAST_PRIMARY_KEYS.set({})
     return result
 
 
@@ -68,13 +70,12 @@ def _delete_cache_entry(
 ) -> None:
     session.delete(cache_entry)
     session.commit()
-
     # Delete cache file
     json.loads(cache_entry._result_as_string, object_hook=clean._delete_cache_file)
 
 
 def hexdigestify_python_call(
-    func: Union[str, Callable[..., Any]],
+    func_to_hex: Union[str, Callable[..., Any]],
     *args: Any,
     **kwargs: Any,
 ) -> str:
@@ -82,7 +83,7 @@ def hexdigestify_python_call(
 
     Parameters
     ----------
-    func: str, callable
+    func_to_hex: str, callable
         Function to hexdigestify
     *args: Any
         Arguments of ``func``
@@ -93,25 +94,39 @@ def hexdigestify_python_call(
     -------
     str
     """
-    return utils.hexdigestify(encode.dumps_python_call(func, *args, **kwargs))
+    return utils.hexdigestify(encode.dumps_python_call(func_to_hex, *args, **kwargs))
 
 
 def cacheable(func: F) -> F:
-    """Make a function cacheable."""
+    """Make a function cacheable.
+
+    The __context__ argument allows to set the `contextvars.Context`.
+    __context__ is not passed to the wrapped function.
+    """
 
     @functools.wraps(func)
-    def wrapper(*args: Any, **kwargs: Any) -> Any:
+    def wrapper(
+        *args: Any, __context__: Optional[contextvars.Context] = None, **kwargs: Any
+    ) -> Any:
+        if __context__:
+            for key, value in __context__.items():
+                key.set(value)
+
+        settings = config.SETTINGS.get()
+        tag = settings.tag
+        expiration = (
+            datetime.datetime.fromisoformat(settings.expiration)
+            if settings.expiration is not None
+            else settings.expiration
+        )
+
         # Cache opt-out
-        if not config.SETTINGS["use_cache"]:
+        if not settings.use_cache:
             return _clear_last_primary_keys_and_return(func(*args, **kwargs))
 
-        # Key defining the function and its arguments
         try:
-            hexdigest = hexdigestify_python_call(
-                func,
-                *args,
-                **kwargs,
-            )
+            # Get key
+            hexdigest = hexdigestify_python_call(func, *args, **kwargs)
         except encode.EncodeError as ex:
             warnings.warn(f"can NOT encode python call: {ex!r}", UserWarning)
             return _clear_last_primary_keys_and_return(func(*args, **kwargs))
@@ -121,72 +136,62 @@ def cacheable(func: F) -> F:
             config.CacheEntry.key == hexdigest,
             config.CacheEntry.expiration > datetime.datetime.utcnow(),
         ]
-        if config.SETTINGS["expiration"]:
+        if expiration is not None:
             # If expiration is provided, only get entries with matching expiration
-            filters.append(
-                config.CacheEntry.expiration
-                == datetime.datetime.fromisoformat(config.SETTINGS["expiration"])
-            )
-        with sqlalchemy.orm.Session(
-            config.SETTINGS["engine"], autoflush=False
-        ) as session:
+            filters.append(config.CacheEntry.expiration == expiration)
+        with sqlalchemy.orm.Session(config.ENGINE.get(), autoflush=False) as session:
             for cache_entry in (
                 session.query(config.CacheEntry)
                 .filter(*filters)
                 .order_by(config.CacheEntry.timestamp.desc())
             ):
+                # Attempt all valid cache entries
                 try:
-                    return _update_last_primary_keys_and_return(session, cache_entry)
+                    return _update_last_primary_keys_and_return(
+                        session, cache_entry, tag
+                    )
                 except decode.DecodeError as ex:
                     # Something wrong, e.g. cached files are corrupted
                     warnings.warn(str(ex), UserWarning)
                     _delete_cache_entry(session, cache_entry)
 
             # Not in the cache
+            cache_entry = None
             try:
-                if _LOCKER:
-                    # Lock cache entry
-                    cache_entry = config.CacheEntry(
-                        key=hexdigest,
-                        expiration=config.SETTINGS["expiration"],
-                        result=_LOCKER,
-                        tag=config.SETTINGS["tag"],
-                    )
-                    session.add(cache_entry)
-                    session.commit()
+                # Lock cache entry
+                cache_entry = config.CacheEntry(
+                    key=hexdigest,
+                    expiration=expiration,
+                    result=_LOCKER,
+                    tag=settings.tag,
+                )
+                session.add(cache_entry)
+                session.commit()
             except sqlalchemy.exc.IntegrityError:
-                # Concurrent job: This cache entry already exist.
+                # Concurrent job: This cache entry already exists.
                 filters = [
                     config.CacheEntry.key == cache_entry.key,
                     config.CacheEntry.expiration == cache_entry.expiration,
                 ]
                 session.rollback()
                 cache_entry = session.query(config.CacheEntry).filter(*filters).one()
-                return _update_last_primary_keys_and_return(session, cache_entry)
-
-            # Compute result from scratch and unlock
-            result = func(*args, **kwargs)
-            try:
-                json_result = json.loads(encode.dumps(result))
-                if _LOCKER:
-                    cache_entry.result = json_result
-                else:
-                    cache_entry = config.CacheEntry(
-                        key=hexdigest,
-                        expiration=config.SETTINGS["expiration"],
-                        result=json_result,
-                        tag=config.SETTINGS["tag"],
-                        counter=0,
+                return _update_last_primary_keys_and_return(session, cache_entry, tag)
+            else:
+                # Compute result from scratch
+                result = func(*args, **kwargs)
+                try:
+                    # Update cache
+                    cache_entry.result = json.loads(encode.dumps(result))
+                    return _update_last_primary_keys_and_return(
+                        session, cache_entry, tag
                     )
-                    session.add(cache_entry)
-                    session.commit()
-                return _update_last_primary_keys_and_return(session, cache_entry)
-            except Exception as ex:
-                if _LOCKER:
+                except encode.EncodeError as ex:
+                    # Enconding error, return result without caching
+                    warnings.warn(f"can NOT encode output: {ex!r}", UserWarning)
+                    return _clear_last_primary_keys_and_return(result)
+            finally:
+                # Unlock
+                if cache_entry and cache_entry.result == _LOCKER:
                     _delete_cache_entry(session, cache_entry)
-                if not isinstance(ex, encode.EncodeError):
-                    raise ex
-                warnings.warn(f"can NOT encode output: {ex!r}", UserWarning)
-                return _clear_last_primary_keys_and_return(result)
 
     return cast(F, wrapper)
